fix(dropbox): Fix a bug where deletions from Dropbox were not awaited
Batch deletes from Dropbox (as part of Maintenance) were not awaited properly.
The function was async, without an await.
This resulted in de function exiting before files had been deleted from Dropbox.

Now, the function returns a promise that is only resolved when Dropbox signals the delete is done. And this promise is properly awaited as part of the maintenance function.